requirements.txt
pymupdf 
transformers 
supabase 
sentence-transformers
tqdm 
python-dotenv
langchain
langchain-text-splitters
